requirements.txt
PyQt5>=5.15.0
pyserial>=3.5
pynmea2>=1.18.0
